+++ File/File.h
@@ -20,10 +20,10 @@ class file {
 
 public:
 
-	unordered_map<int, coordinates> readFile(string fileName,  unordered_map<int, vector<edge>>& graph, float& maxSpeed);
-	unordered_map<int, coordinates> readFile(string fileName,unordered_map<int,vector<edge>>&graph,float &maxSpeed, float& Speed_Interval, int& speedSize);
+	unordered_map<int, coordinates> readFile(string fileName,  unordered_map<int, vector<edge>>& graph, double& maxSpeed);
+	unordered_map<int, coordinates> readFile(string fileName,unordered_map<int,vector<edge>>&graph,double &maxSpeed, double& Speed_Interval, int& speedSize);
 	void writeFile(string fileName,vector<bestPath> quries);
-	string getCellKey(float x, float y, float cellSize);
+	string getCellKey(double x, double y, double cellSize);
 
 	vector<query> readQuery(string fileName);
